+++ core/impl/src/main/java/com/blazebit/persistence/impl/FunctionalDependencyAnalyzerVisitor.java
@@ -48,11 +48,15 @@ import com.blazebit.persistence.parser.util.JpaMetamodelUtils;
 import com.blazebit.persistence.spi.ExtendedManagedType;
 
 import javax.persistence.metamodel.Attribute;
+import javax.persistence.metamodel.PluralAttribute;
 import javax.persistence.metamodel.SingularAttribute;
+import javax.persistence.metamodel.Type;
 import java.util.AbstractMap;
 import java.util.ArrayList;
+import java.util.Collections;
 import java.util.HashMap;
 import java.util.HashSet;
+import java.util.LinkedHashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.Set;
@@ -62,49 +66,101 @@ import java.util.Set;
  * @author Christian Beikov
  * @since 1.3.0
  */
-class UniquenessDetectionVisitor extends AbortableVisitorAdapter {
+class FunctionalDependencyAnalyzerVisitor extends AbortableVisitorAdapter {
 
     private final EntityMetamodel metamodel;
     private final Map<Object, Set<String>> uniquenessMissingJoinNodeAttributes;
     private final Map<Object, List<ResolvedExpression>> uniquenessFormingJoinNodeExpressions;
+    private final Map<Object, List<ResolvedExpression>> functionalDependencyRootExpressions;
     private Object lastJoinNode;
     private ResolvedExpression currentResolvedExpression;
     private boolean resultUnique;
+    private boolean inKey;
 
-    public UniquenessDetectionVisitor(EntityMetamodel metamodel) {
+    public FunctionalDependencyAnalyzerVisitor(EntityMetamodel metamodel) {
         this.metamodel = metamodel;
         this.uniquenessMissingJoinNodeAttributes = new HashMap<>();
         this.uniquenessFormingJoinNodeExpressions = new HashMap<>();
+        this.functionalDependencyRootExpressions = new LinkedHashMap<>();
     }
 
     public void clear() {
         uniquenessMissingJoinNodeAttributes.clear();
         uniquenessFormingJoinNodeExpressions.clear();
+        functionalDependencyRootExpressions.clear();
         resultUnique = false;
+        lastJoinNode = null;
     }
 
-    public boolean isUnique(Expression expression) {
+    public boolean analyzeFormsUniqueTuple(Expression expression) {
         lastJoinNode = null;
         boolean unique = expression.accept(this);
+        JoinNode p;
+        if (lastJoinNode instanceof Map.Entry<?, ?>) {
+            p = ((Map.Entry<JoinNode, ?>) lastJoinNode).getKey();
+        } else {
+            p = (JoinNode) lastJoinNode;
+        }
+        if (p != null) {
+            // Traverse join nodes up and if we see that this expression is part of a collection, we revert result uniqueness
+            while (p.getParent() != null) {
+                if (p.getParentTreeNode() == null || p.getParentTreeNode().isCollection()) {
+                    unique = false;
+                    break;
+                }
+                p = p.getParent();
+            }
+        }
         resultUnique = resultUnique || unique;
         return unique;
     }
 
-    public ResolvedExpression[] getResultUniqueExpressions(ResolvedExpression[] expressions) {
+    public ResolvedExpression[] getFunctionalDependencyRootExpressions(ResolvedExpression[] expressions) {
         clear();
         for (int i = 0; i < expressions.length; i++) {
             currentResolvedExpression = expressions[i];
-            isUnique(expressions[i].getExpression());
-            currentResolvedExpression = null;
-            if (isResultUnique()) {
+            if (analyzeFormsUniqueTuple(expressions[i].getExpression())) {
                 List<ResolvedExpression> resolvedExpressions = uniquenessFormingJoinNodeExpressions.get(lastJoinNode);
-                lastJoinNode = null;
-                return resolvedExpressions.toArray(new ResolvedExpression[resolvedExpressions.size()]);
+                if (resolvedExpressions != null) {
+                    functionalDependencyRootExpressions.put(lastJoinNode, resolvedExpressions);
+                    lastJoinNode = null;
+                }
             }
         }
 
+        currentResolvedExpression = null;
         lastJoinNode = null;
-        return null;
+        if (functionalDependencyRootExpressions.isEmpty()) {
+            return null;
+        }
+        List<ResolvedExpression> resolvedExpressions;
+        if (functionalDependencyRootExpressions.size() == 1) {
+            resolvedExpressions = functionalDependencyRootExpressions.values().iterator().next();
+        } else {
+            resolvedExpressions = new ArrayList<>();
+            OUTER: for (Map.Entry<Object, List<ResolvedExpression>> entry : functionalDependencyRootExpressions.entrySet()) {
+                JoinNode node;
+                if (entry.getKey() instanceof Map.Entry<?, ?>) {
+                    node = ((Map.Entry<JoinNode, ?>) entry.getKey()).getKey();
+                } else {
+                    node = (JoinNode) entry.getKey();
+                }
+                // We always need collections in here
+                if (node.getParentTreeNode() == null || node.getParentTreeNode().isCollection()) {
+                    resolvedExpressions.addAll(entry.getValue());
+                } else {
+                    // Skip *ToOne join nodes that are functionally dependent on an existing node
+                    while (node.getParent() != null) {
+                        if (functionalDependencyRootExpressions.containsKey(node.getParent())) {
+                            continue OUTER;
+                        }
+                        node = node.getParent();
+                    }
+                    resolvedExpressions.addAll(entry.getValue());
+                }
+            }
+        }
+        return resolvedExpressions.toArray(new ResolvedExpression[resolvedExpressions.size()]);
     }
 
     public boolean isResultUnique() {
@@ -113,18 +169,39 @@ class UniquenessDetectionVisitor extends AbortableVisitorAdapter {
 
     @Override
     public Boolean visit(PathExpression expr) {
-        if (expr.getField() == null) {
+        PathReference pathReference = expr.getPathReference();
+        JoinNode baseNode = (JoinNode) pathReference.getBaseNode();
+
+        if (pathReference.getField() == null) {
+            lastJoinNode = baseNode;
+            functionalDependencyRootExpressions.put(baseNode, Collections.singletonList(currentResolvedExpression));
+            // This is a basic element collection. The element is it's unique key
+            if (baseNode.getType().getPersistenceType() == Type.PersistenceType.BASIC) {
+                return true;
+            }
+            // The key of a collection is it's unique key
+            if (inKey) {
+                return true;
+            }
             throw new IllegalArgumentException("Ordering by association '" + expr + "' does not make sense! Please order by it's id instead!");
         }
 
         // First we check if the target attribute is unique, if it isn't, we don't need to check the join structure
-        PathReference pathReference = expr.getPathReference();
-        JoinNode baseNode = (JoinNode) pathReference.getBaseNode();
         ExtendedManagedType<?> managedType = metamodel.getManagedType(ExtendedManagedType.class, baseNode.getJavaType());
         Attribute attr = managedType.getAttribute(pathReference.getField()).getAttribute();
 
+        if (attr instanceof PluralAttribute<?, ?, ?>) {
+            lastJoinNode = baseNode;
+            if (inKey) {
+                registerFunctionalDependencyRootExpression(baseNode);
+                return true;
+            }
+            throw new IllegalArgumentException("Ordering by plural attribute '" + expr + "' does not make sense! Please order by it's id instead!");
+        }
+
         // Right now we only support ids, but we actually should check for unique constraints
         if (!((SingularAttribute<?, ?>) attr).isId()) {
+            registerFunctionalDependencyRootExpression(baseNode);
             return false;
         }
 
@@ -137,12 +214,15 @@ class UniquenessDetectionVisitor extends AbortableVisitorAdapter {
             // We have to correct the base node for single valued id paths
             String associationName = expr.getField().substring(0, dotIndex);
             Attribute<?, ?> attribute = baseNode.getManagedType().getAttribute(associationName);
+            baseNodeKey = new AbstractMap.SimpleEntry<>(baseNode, associationName);
             if (attribute.getPersistentAttributeType() != Attribute.PersistentAttributeType.ONE_TO_ONE) {
+                registerFunctionalDependencyRootExpression(baseNodeKey);
                 return false;
             }
-            baseNodeKey = new AbstractMap.SimpleEntry<>(baseNode, associationName);
             managedType = metamodel.getManagedType(ExtendedManagedType.class, JpaMetamodelUtils.resolveFieldClass(baseNode.getJavaType(), attribute));
         }
+
+        registerFunctionalDependencyRootExpression(baseNodeKey);
 
         // First we initialize the names of the id attributes as set for the join node
         Set<String> orderedAttributes = uniquenessMissingJoinNodeAttributes.get(baseNodeKey);
@@ -193,6 +273,39 @@ class UniquenessDetectionVisitor extends AbortableVisitorAdapter {
         }
 
         return true;
+    }
+
+    private void registerFunctionalDependencyRootExpression(Object baseNode) {
+        if (!uniquenessFormingJoinNodeExpressions.containsKey(baseNode)) {
+            List<ResolvedExpression> resolvedExpressions = functionalDependencyRootExpressions.get(baseNode);
+            if (resolvedExpressions == null) {
+                resolvedExpressions = new ArrayList<>();
+                functionalDependencyRootExpressions.put(baseNode, resolvedExpressions);
+            }
+            resolvedExpressions.add(currentResolvedExpression);
+        }
+    }
+
+    @Override
+    public Boolean visit(ListIndexExpression expression) {
+        boolean oldInKey = inKey;
+        inKey = true;
+        try {
+            return expression.getPath().accept(this);
+        } finally {
+            inKey = oldInKey;
+        }
+    }
+
+    @Override
+    public Boolean visit(MapKeyExpression expression) {
+        boolean oldInKey = inKey;
+        inKey = true;
+        try {
+            return expression.getPath().accept(this);
+        } finally {
+            inKey = oldInKey;
+        }
     }
 
     @Override
@@ -250,17 +363,7 @@ class UniquenessDetectionVisitor extends AbortableVisitorAdapter {
     }
 
     @Override
-    public Boolean visit(ListIndexExpression expression) {
-        return false;
-    }
-
-    @Override
     public Boolean visit(MapEntryExpression expression) {
-        return false;
-    }
-
-    @Override
-    public Boolean visit(MapKeyExpression expression) {
         return false;
     }